Added optional value to flutter_flow_choice_chips.dart
Added option value behind the label. This is a very common usecase for these multichoice widgets especially when working with multiple languages. The label displayed to the user will often be different than what we store as the value. 

For example, when you have a list of enum a user can select - you will translate the enum to string but keep the enum as the value you store